requirements.txt
aiogram==2.23.1
aiohttp==3.8.0
aiosignal==1.3.1
APScheduler==3.11.0
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
certifi==2024.8.30
chardet==3.0.4
charset-normalizer==2.1.1
et_xmlfile==2.0.0
frozenlist==1.5.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
magic-filter==1.0.12
multidict==6.1.0
numpy==2.1.3
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0
propcache==0.2.0
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
TgCrypto==1.2.5
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
yarl==1.18.0
